requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn
scipy

# Deep Learning & Advanced ML (for future extensions)
# torch==2.0.1
# tensorflow==2.13.0

# API Framework
fastapi
uvicorn[standard]==0.23.2
pydantic
pydantic-settings==2.0.3

# Workflow Orchestration
prefect
prefect-aws==0.3.0

# ML Testing & Validation
deepchecks

# Data Processing
python-multipart==0.0.6

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2

# Frontend
streamlit==1.28.0
matplotlib==3.8.0
requests==2.31.0

# Optional: For notifications (Slack/Discord placeholders)
# slack-sdk==3.21.3
# discord.py==2.3.1

